+++ Domain/DTO/DroneDispatchCommunication/AssignDeliveryRequest.cs
@@ -1,5 +1,4 @@
 ﻿using System.Text.Json.Serialization;
-using Domain.DTO.Shared;
 using Domain.Entities;
 using MongoDB.Bson.Serialization.Attributes;
