+++ Domain/DTO/DroneState.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain
+﻿namespace Domain.DTO
 {
     public class DroneState
     {
