+++ Domain/Entities/Order.cs
@@ -2,7 +2,6 @@
 using System.Linq;
 using System.Text.Json.Serialization;
 using Domain.DTO;
-using Domain.DTO.Shared;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
